icmr-parser/requirements.txt
pymupdf==1.24.0
pdfplumber==0.11.0
groq==0.9.0
langchain==0.3.0
langchain-groq==0.2.0
pydantic==2.7.0
pandas==2.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.0
google-genai>=0.8.0
httpx>=0.25.0

